pyproject.toml
[tool.poetry]
name = "dlp"
version = "0.1.0"
description = "Test task for Truely"
authors = ["Stepan Pletnev <spletnev@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ruff = "^0.3.2"
flask-admin = "^1.6.1"
flask = "^3.0.2"
independency = "^1.2"
pydantic = "^2.6.3"
pytest = "^8.1.1"
fastapi = "^0.110.0"
pydantic-settings = "^2.2.1"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
celery = {extras = ["redis"], version = "^5.3.6"}
slack-bolt = "^1.18.1"
uvicorn = "^0.28.0"
flask-sqlalchemy = "^3.1.1"
walrus = "^0.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
ignore-init-module-imports = true
exclude = ["__init__.py"]